refactor: modularize scripts and add bundler
