+++ package.json
@@ -1,10 +1,13 @@
 {
   "name": "contucuota",
   "version": "1.0.0",
+  "type": "module",
   "scripts": {
-    "test": "jest"
+    "test": "jest",
+    "build": "vite build"
   },
   "devDependencies": {
-    "jest": "^29.7.0"
+    "jest": "^29.7.0",
+    "vite": "^5.2.0"
   }
 }
