+++ plataforma.html
@@ -29,7 +29,7 @@
   <link rel="stylesheet" href="https://unpkg.com/aos@2.3.1/dist/aos.css" />
   <!-- Carga diferida de scripts no críticos -->
   <script defer src="https://unpkg.com/aos@2.3.1/dist/aos.js"></script>
-  <script defer src="./plataforma.js"></script>
+  <script type="module" defer src="./dist/plataforma.bundle.js"></script>
 </head>
 <body>
   <!-- Enlace de accesibilidad para saltar al contenido principal -->
