+++ simulador.html
@@ -29,7 +29,7 @@
   <!-- Carga diferida de scripts no críticos -->
   <script defer src="https://cdn.jsdelivr.net/npm/chart.js"></script>
   <script defer src="https://unpkg.com/aos@2.3.1/dist/aos.js"></script>
-  <script defer src="./simulador.js"></script>
+  <script type="module" defer src="./dist/simulador.bundle.js"></script>
 </head>
 <body>
   <!-- Enlace de accesibilidad para saltar al contenido principal -->
